// Assets-Uncapped/front-portal-develop/src/services/api/organisation-users/models/DocumentByTypeResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * api-organisation-users
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: @project.version@
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { MissingDocumentResponse } from './MissingDocumentResponse';
import {
    MissingDocumentResponseFromJSON,
    MissingDocumentResponseFromJSONTyped,
    MissingDocumentResponseToJSON,
    MissingDocumentResponseToJSONTyped,
} from './MissingDocumentResponse';
import type { UploadedDocumentResponse } from './UploadedDocumentResponse';
import {
    UploadedDocumentResponseFromJSON,
    UploadedDocumentResponseFromJSONTyped,
    UploadedDocumentResponseToJSON,
    UploadedDocumentResponseToJSONTyped,
} from './UploadedDocumentResponse';

/**
 * 
 * @export
 * @interface DocumentByTypeResponse
 */
export interface DocumentByTypeResponse {
    /**
     * 
     * @type {Array<UploadedDocumentResponse>}
     * @memberof DocumentByTypeResponse
     */
    uploadedDocuments?: Array<UploadedDocumentResponse>;
    /**
     * 
     * @type {MissingDocumentResponse}
     * @memberof DocumentByTypeResponse
     */
    requiredDocument?: MissingDocumentResponse;
}

/**
 * Check if a given object implements the DocumentByTypeResponse interface.
 */
export function instanceOfDocumentByTypeResponse(value: object): value is DocumentByTypeResponse {
    return true;
}

export function DocumentByTypeResponseFromJSON(json: any): DocumentByTypeResponse {
    return DocumentByTypeResponseFromJSONTyped(json, false);
}

export function DocumentByTypeResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): DocumentByTypeResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'uploadedDocuments': json['uploadedDocuments'] == null ? undefined : ((json['uploadedDocuments'] as Array<any>).map(UploadedDocumentResponseFromJSON)),
        'requiredDocument': json['requiredDocument'] == null ? undefined : MissingDocumentResponseFromJSON(json['requiredDocument']),
    };
}

export function DocumentByTypeResponseToJSON(json: any): DocumentByTypeResponse {
    return DocumentByTypeResponseToJSONTyped(json, false);
}

export function DocumentByTypeResponseToJSONTyped(value?: DocumentByTypeResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'uploadedDocuments': value['uploadedDocuments'] == null ? undefined : ((value['uploadedDocuments'] as Array<any>).map(UploadedDocumentResponseToJSON)),
        'requiredDocument': MissingDocumentResponseToJSON(value['requiredDocument']),
    };
}