// Assets-Uncapped/front-portal-develop/src/services/api/partners/models/UploadedDocumentResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * api-partners
 * Change log: <br>1.0.0 initial release
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface UploadedDocumentResponse
 */
export interface UploadedDocumentResponse {
    /**
     * 
     * @type {string}
     * @memberof UploadedDocumentResponse
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof UploadedDocumentResponse
     */
    fileName?: string;
    /**
     * 
     * @type {number}
     * @memberof UploadedDocumentResponse
     */
    fileSize?: number;
    /**
     * 
     * @type {string}
     * @memberof UploadedDocumentResponse
     */
    documentType?: string;
    /**
     * 
     * @type {string}
     * @memberof UploadedDocumentResponse
     */
    status?: string;
    /**
     * 
     * @type {Date}
     * @memberof UploadedDocumentResponse
     */
    updatedAt?: Date;
    /**
     * 
     * @type {string}
     * @memberof UploadedDocumentResponse
     */
    errorMessage?: string;
    /**
     * 
     * @type {string}
     * @memberof UploadedDocumentResponse
     */
    requiredDocumentId?: string;
}

/**
 * Check if a given object implements the UploadedDocumentResponse interface.
 */
export function instanceOfUploadedDocumentResponse(value: object): value is UploadedDocumentResponse {
    return true;
}

export function UploadedDocumentResponseFromJSON(json: any): UploadedDocumentResponse {
    return UploadedDocumentResponseFromJSONTyped(json, false);
}

export function UploadedDocumentResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): UploadedDocumentResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'fileName': json['fileName'] == null ? undefined : json['fileName'],
        'fileSize': json['fileSize'] == null ? undefined : json['fileSize'],
        'documentType': json['documentType'] == null ? undefined : json['documentType'],
        'status': json['status'] == null ? undefined : json['status'],
        'updatedAt': json['updatedAt'] == null ? undefined : (new Date(json['updatedAt'])),
        'errorMessage': json['errorMessage'] == null ? undefined : json['errorMessage'],
        'requiredDocumentId': json['requiredDocumentId'] == null ? undefined : json['requiredDocumentId'],
    };
}

export function UploadedDocumentResponseToJSON(json: any): UploadedDocumentResponse {
    return UploadedDocumentResponseToJSONTyped(json, false);
}

export function UploadedDocumentResponseToJSONTyped(value?: UploadedDocumentResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'fileName': value['fileName'],
        'fileSize': value['fileSize'],
        'documentType': value['documentType'],
        'status': value['status'],
        'updatedAt': value['updatedAt'] == null ? undefined : ((value['updatedAt']).toISOString()),
        'errorMessage': value['errorMessage'],
        'requiredDocumentId': value['requiredDocumentId'],
    };
}